move content from rcl_utils.rs to arguments.rs
